Extract version management from Gradle scripts
The release versions are now extracted and incremented based on the most
recent release tag in git.

This change also adds convenience methods for triggering a new release.
In practice these methods are Gradle tasks that will create and push a
new tag to the remote servers. This new tag will in turn trigger Circle
CI to initialize the deployment flow.