Allows the mock web server to consume the request body
This change allows the MockWebServer to consume the request body before
asking for a mock response. The mock response provider is then allowed
to read from the consumed buffer instead of the socket input stream
directly.

An alternative would be to expose the socket input stream to the mock
response provider, but that would require guarantees that it will fully
exhausts the input stream every time, regardless if it needs it or not.
Failing in doing so would cause truncating issues when reading the next
request from the input stream.

The conclusion is that there's no need to couple the two infrastructures
that tightly. Keeping the input stream exhaustion responsibility in the
mock web server seems the most reasonable thing to do as of now.