Allows MockRequest being created from Meta
This is an internal helper feature and it's important while the Meta
class isn't exposed to the public (and shouldn't be).

An upcoming feature of Atlantis is to expose the conversion between
mocked notation and "real world" notation (i.e. switching from
"http://localhost:8080" to "https://api.example.com"). Given this, we
still don't want to expose the Meta class but instead use the already
public MockRequest class.