Moves and renames StateHelper
StateHelper was a confusing name that didn't reflect what the API really
is supposed to do. The new name; "TokenHelper" is considered a better
name.

Also the feature is moved to the top level Atlantis domain in order to
be applicable to all requests and responses (even those that are relayed
to a real server and, hence, doesn't have a configuration yet).